Move all Insights related gql queries in one place
  Fix updatedAt to publishedAt for proper calculation of VisualBacktestChart
