+++ src/components/Insight/InsightsFeatured.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { graphql } from 'react-apollo'
 import InsightCardSmall from './InsightCardSmall'
-import { FEATURED_INSIGHTS_QUERY } from './insightsGQL'
+import { FEATURED_INSIGHTS_QUERY } from './../../queries/InsightsGQL'
 import { creationDateSort } from '../../pages/Insights/utils'
 
 const InsightsTrends = ({ data: { insights = [] }, ...props }) => {
