--- src/components/Insight/insightsGQL.js
@@ -1,83 +1,0 @@
-import gql from 'graphql-tag'
-
-export const FEATURED_INSIGHTS_QUERY = gql`
-  query featuredInsights {
-    insights: featuredInsights {
-      id
-      title
-      createdAt
-      votedAt
-      votes {
-        totalVotes
-      }
-      tags {
-        name
-      }
-      user {
-        id
-        username
-      }
-    }
-  }
-`
-
-export const ALL_INSIGHTS_BY_TAG_QUERY = gql`
-  query allInsightsByTag($tag: String!) {
-    allInsightsByTag(tag: $tag) {
-      readyState
-      id
-      title
-      createdAt
-      votedAt
-      votes {
-        totalVotes
-      }
-      tags {
-        name
-      }
-      user {
-        id
-        username
-      }
-    }
-  }
-`
-
-export const ALL_INSIGHTS_QUERY = gql`
-  query allInsights {
-    allInsights {
-      readyState
-      id
-      title
-      createdAt
-      votedAt
-      votes {
-        totalVotes
-      }
-      tags {
-        name
-      }
-      user {
-        id
-        username
-      }
-    }
-  }
-`
-
-export const PINNED_INSIGHTS_QUERY = gql`
-  query post($id: Int!) {
-    pinnedInsight: post(id: $id) {
-      id
-      title
-      createdAt
-      tags {
-        name
-      }
-      user {
-        id
-        username
-      }
-    }
-  }
-`
