+++ src/pages/Insights/insightDraftEpic.js
@@ -4,7 +4,7 @@ import {
   CREATE_INSIGHT_DRAFT_MUTATION,
   UPDATE_INSIGHT_DRAFT_MUTATION,
   PUBLISH_INSIGHT_DRAFT_MUTATION
-} from './InsightsGQL'
+} from './../../queries/InsightsGQL'
 import * as actions from './actions'
 import { getInsightTrendTagByDate } from '../../components/Insight/InsightsTrends'
 import { completeOnboardingTask } from '../../pages/Dashboard/utils'
